Update patch set 2
Patch Set 2:

Hi Moiz, you are setting the name only for static flows (The ones configured through FlowConfig objects). This change has no effects for all the flows configured by osgi applications (like SimpleForwarding).

Again, the Flow class is to represent the flow to be installed in hardware. No name concept here.
I am still for changing the Flow.id type to String so that is more generic.
From northbound this is a backward compatible change. The only compatibility issue is for saved to file static flows.

But considering the cookie field is likely not to be set by an app, I am for changing its type in Flow.

I can push a change for this now.

Patch-set: 2