Update patch set 2
Patch Set 2:

Alessandro, I see that you changed the id to a string. I am still not sure how that would remove the need to pass for me to set a name (I understand that flows in of do not have a name) but we do need to use the name in md-sal as the id. Do you see any alternative to what I can do?

Patch-set: 2
Reviewer: Gerrit User 795 <795@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0