Modify uncaught exception handler to call System.exit(1).
This introduces the ability to specify how uncaught exceptions are
handled via setting controller.uncaughtExceptionPolicy property.

There are two policies: IGNORE and ABORT. The former logs which thread
has died and does nothing else. The latter forces a virtual machine
shutdown when a thread dies.

The default policy is IGNORE, as that has been the historic setting. For
Helium release, this should probably get discussed, as uncontrolled
thread death may leave the controller in an inconsistent/invalid state.

Change-Id: Ib7c64b23e1cd3c7c45cfa31401410bba22acd431
Signed-off-by: Tomas Olvecky <tolvecky@cisco.com>
Signed-off-by: Robert Varga <rovarga@cisco.com>