Make MD-SAL use opendaylight parent and unify model versions
This commit aligns the various model versions used under MD-SAL
component with the rest of the project by hooking MD-SAL parent to
the common parent.

Since MD-SAL code does not comply with coding style enforced by the
checkstyle plugin, disable its execution for now -- BUG-272 tracks the
need to re-enable it.

Bumped version of Javassist to 3.18.1-GA which is required by PowerMock
and MD-SAL.

Change-Id: I55763136cec2ce0e892a3c7cc6543f913d76a2ad
Signed-off-by: Robert Varga <rovarga@cisco.com>
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>