Fixing a NullPointerException in Neutron Interface (https://bugs.opendaylight.org/show_bug.cgi?id=323)
Ideally, this condition should never occur. But there seems to be a consistency issue in the ML2 driver
which could potentially result in this condition. While we are addressing the root-cause in ODL ML2
driver, we should also patch our side of the interface as well to avoid failing for such conditions when
it is harmless for a few usecases.

Change-Id: Iff5fdb76b45280aba38607cdcf82715cb4338b05
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>