Update patch set 1
Patch Set 1: Code-Review-1

This sysouts are there only for convenience of debuging
The templates you modified are for runtime generated code.

In case of runtime generated code and OSGI when you are extending other bundles you must do that with great care,
What if classloader does not have access to SLF4J, etc.
Also you changed that to log field, which is defined in the template,
but this field will not be propagated to the resulting new classes,
which if this lines are uncommented, will result in compilation failure
at runtime.

Patch-set: 1
Reviewer: Gerrit User 21 <21@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1