Update patch set 1
Patch Set 1:

Yes it is true that we already have a inventory-notification which gets the notification and updates the Operational DataStore.

There was a use-case, if some some north-bound module registers for the same notification and tries to read the data from the Operational DataStore, then they will face a null pointer exception. 
The problem is when both inventory-manager and northbound module registers for the same notification, but inventory-manager writes the data while other module reads the same data from the dataStore, both the notifications are processed by seperate threads, hence its possible that the north-bound module notification is processed first, in that case it will face a null data problem. 

In order to overcome the problem, a new notification is published when the data is available in the DataStore. So If a application wants to read the data from the DataStore then they can register for the new notification, or in othercase they can register for the existing notification.

Patch-set: 1
Reviewer: Gerrit User 1471 <1471@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0