Handle uncaught exceptions from Clustering Services
There are couple of timer tasks which use caches provided by
clustening services. If a container is deleted, then the cache
allocated to that container is also removed. Although TimerTasks
do get cancelled when container is deleted, but if happens during
the execution of timer thread, then the task may try to access a
cache which has already been deallocated.

The accessing of this deallocated cache will result in clustering
service exception which needs to be handled by timer tasks

Change-Id: Id1c481360a4ff71b6eb0afab3c030ef011ca7dcd
Signed-off-by: Maurice Qureshi <maquresh@cisco.com>