Removed replacement of specific characters (e.g. :) in key part of URI
Key part of URI in
nodes/node/openflow:1
is openflow:1. It was encoded like
nodes/node/openflow%3A1 and used as stream ID when listener was created
but when such URI was used as identificator of stream it wasn't posible to subscribe to this stream via calling
streams/stream/nodes/node/openflow%3A1
because input parameter of method subscribeToStream - identifier -
contained URI in first form (incorrect ID) and therefore it wasn't
possible to find listener for stream.

Change-Id: I10fa4247581f933e014f889c529616289082bdb6
Signed-off-by: Jozef Gloncak <jgloncak@cisco.com>