DO NOT MERGE : Use a DataChange listener instead of handling NodeConnectorUpdated events in the compatibility layer
This was done to prevent a race condition in the NodeConnectorUpdated handler where we were looking up the data in the operational
tree to figure out if the node connector was added or updated and many times finding the data in there and assuming that it was a
change when it was an add.

To address the issue now we participate in a transaction such that whenever a node connector is modified the NodeConnectorCommitHandler
is invoked and there we figure out if the NodeConnector was added or removed.

Using the Data change listener does not however solve the problem. In fact it makes it worse and I would like this code to be examined to
see if it's because I am doing something wrong.

One more thing I wanted to try was to set up Data Change listeners for Nodes as well however that does not seem to be working very well because
in my NodeCommitHandler the data maps sometimes contain objects other than a Node (for example many times they contain a table). This could possibly be
because my data path is not right. Also something to be looked at.

Change-Id: Ide2e6e85191856a8c6a9746b2397605905938d56
Signed-off-by: Moiz Raja <moraja@cisco.com>