Use a DataChange listener instead of handling NodeConnectorUpdated events in the compatibility layer
While this commit does seem to fix the problem it seems to introduce a new problem. When trying the following sequence you will
end up with the datastore not being cleaned up on mininet exit

1. Start mini net with tree,3 topology
2. ping all
3. exit mini net

Using restconf GET http://localhost:8080/restconf/operational/opendaylight-inventory:nodes

Nodes will still be there. I have tested this with a distribution built in the openflow plugin repo.

Apparently this is a known issue - Tony is fixing this.

Change-Id: Ide2e6e85191856a8c6a9746b2397605905938d56
Signed-off-by: Moiz Raja <moraja@cisco.com>