Add watchdog thread to detect illegal blocking calls during second phase commit.
Add a thread that starts when second phase commit starts, that will monitor if
a module instance occupies the commit thread for more than 5 seconds. If so, log
a warning every second informing about possible misuse of config subsystem to
block on external resources.
Added a proxy module that caches Module#getInstance() call as this method is
idempotent.

Change-Id: Ia1e56386bdd8c6b7a6dc625cfc69253b927c92a9
Signed-off-by: Tomas Olvecky <tolvecky@cisco.com>