Update patch set 2
Patch Set 2:

externalized the repositories into a settings.xml and profile, which means a build needs to specify the settings.xml (-s settings.xml). This does allow the addition of other repositories without having to muck with all the poms and figure out where it needs to go.

Patch-set: 2
Label: Verified=0