Update patch set 1
Patch Set 1:

Thanks for the feedback. I was mostly pushing this is a stopgap to allow people to have "expected" behavior when they disabled SimpleForwarding.

In the medium-term, it probably makes sense to simply move the code that reactively forwards traffic from ArpHandler to SimpleForwarding thus separating concerns a bit better and achieving the same goal.

In the long-term, we're working on a more complete set of learning switch modules which will likely replace all of this.

I'll leave it up you whether to merge this patch or wait for me to at least push the medium-term version.

Patch-set: 1
Reviewer: Gerrit User 48 <48@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0