Bug 510: Updated initial configuration to easily enable new datastore
In order to enable new in-memory datastore (and to wire it with clients)
initial configuration of md-sal (01-md-sal.xml) needs to be updated
to unwire old services and wire new one to the actual MD-SAL brokers.

This could be done by commenting out all DATA-BROKER sections
and uncommenting NEW-DATA-BROKER sections.

Change-Id: I072a65785593b82ec4b1ff6e3dae4c0af7497659
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>
Signed-off-by: Martin Sunal <msunal@cisco.com>
Signed-off-by: Robert Varga <rovarga@cisco.com>