Bug 698 - Confingured controller script to set max memory to 1 GB if not defined
Modified the run.sh and run.bat scripts to configure the java max memory to be
 1 GB if it is not defined on the command line.

NOTE: Java generally sets the memory to be 1/4th the total memory of your system
so it is possible that this will result in LESS memory being allocate (change of
behavior). Once this work item is committed we should encourage a best practice
of defining max memory.

A message will be printed out notifying the users that the max memory was not
specified.

Change-Id: I39225f28c4aed3636a1db0cdb41fbe802f5b5b54
Signed-off-by: dbavery <devin.avery@brocade.com>