Update patch set 1
Patch Set 1:

Hi Ed,

You are correct that eclipse is not doing the source code generation. We had to time that box effort and unfortunetly could not get eclipse generating the yang / xtend source code correctly.

These are the steps that developers have to take to get a clean eclipse build.

1) download the source code
2) run a 'mvn clean install -D skipTests' (technically you only need to run the generate sources phase, but a mvn clean install is likely easier). This generates the yang / xtend sources.
3) Import the projects into eclipse.
       - at this step it seems that the yang-tools project fails the import on the first time. However, right click the project and doing Maven->Update Project and all compile errors should now be gone.

In a future task we can reexamine the yang / xtend plugins for eclipse and see if we can't get the code generating automatically via eclipse.

Patch-set: 1