Update patch set 2
Patch Set 2:

Re: 500's.
Can you clarify what you mean here?
Do you mean that we should let an exception occur and have the proper error page shown with Stack Trace?

Personally I'd rather raise a proper HTTP Error.
I'm thinking that 500 though was a poor choice on my part.

Maybe 404 for "Marker not found". This should only happen if there is a typo in the query parameter or if the network has been deleted.

And maybe 400 for "Marker or limit out of bounds". This is a corner case that *should* never happen, but I was trying to cover every eventuality. E.g when page_reverse=True, marker is at position 0 of the array (or less).

I'll add pagination for Ports and Subnets once I've got a +2 on this one.

Patch-set: 2