Update patch set 2
Patch Set 2:

I am attempting to do a couple of things. The first is to externalize the repository information to a central place as opposed to have it in all the pom files. Having this repeated in various placed, imho, can lead to errors. Having it in a single place it is just easier to manage, at the cost of having to specify the settings file when using maven (people will disagree on the cost / benefit here), but it is really not more burdensome then also requiring people to set the memory limits for the build.

The second thing I am looking at when externalizing the repositories is the ability to add custom repositories into the build in a single place. For example, say your company won't proxy to another nexus repository, with this settings file that can be added in. This can also be used to pull all ODL artifacts from a customer nexus instance as well as opposed from the standard ODL.

Patch-set: 2