Update patch set 2
Patch Set 2: Code-Review+1

I strongly support this patch. The restconf connector bundle as is should not be dependent on a particular implementation, CORS or any other HTTP-level policy needs to be contained within the HTTP endpoint definition.

If there are no further objections, I kindly ask for this patch to be merged as soon as possible.

Patch-set: 2
Label: Code-Review=+1