Update patch set 2
Patch Set 2:

https://code.google.com/p/sortpom/

Add this plugin to any pom and it will sort it according to configuration :-)

At first blush this seems to have much of the sorting that we're looking for.  Not sure we want to have it modify the pom.xml in place but I haven't thought that through.

Patch-set: 2
Reviewer: Gerrit User 1569 <1569@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0