Update patch set 1
Patch Set 1:

I would prefer to not start the TCP server at all if it's not specified. The point is that the config pusher no longer requires it, so we should be able to complete startup without instantiating it -- which I think will be useful for minimal distributions.

Patch-set: 1
Label: Code-Review=0