Bug 735 - Part 2 - modify the controller to reference the ietf-restconf which was added in part 1.
modify the controller to ship the restconf bundle (yang file) so it can be picked up by the controller MD-SAL.

NOTE: This includes a regression back to the version of rest conf defined on 2013-09-04 because ODL does not provide the updated yang types required by the Oct 19 version of the draft.

NOTE: This is dependent on https://git.opendaylight.org/gerrit/#/c/6182 delivered to the yangtools project. This was merged on 4/16

Change-Id: If95e2707199f6969b7ca9fb3366f19309271bef3
Signed-off-by: Devin Avery <devin.avery@brocade.com>