BUG 652 leafref CCE & BUG 720 colons problem
BUG 652 leafref Class cast exception
By default if input data were in format which was possible to parse as
XPath if was done like this. It caused that for leafref which indexed to
instance-identifier the value was parsed as XPath and not as string.

BUG 720 colons problem in JSON
If input string contained colon (:) then it was dealed as namespace:value
and object of type IdentityValuesDTO was created (it is probable that
data could be of type identityref). If data wasn't of type leafref then it
was incorrectly processed as leafref.
The RestCodec was updated to correctly translate data (cases when data
contains ':' but aren't of type leafref)

Change-Id: I0cba06bce1b1c69f6901cc90d2b30a3735e5f57e
Signed-off-by: Jozef Gloncak <jgloncak@cisco.com>