Update patch set 1
Patch Set 1: Code-Review-1

In the integration test bundle the unit tests is referring to the KitchenService which we since removed. Just need to update the reference to KitchenServiceImpl.

There also seems to be some merges with main that need to be resolved now.

Patch-set: 1
Reviewer: Gerrit User 1840 <1840@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1