Update patch set 5
Patch Set 5:

We need to use properties for versions in the feature file.

The feature 'as is' installs. I see:

2014-05-07 10:55:19,561 | WARN  | Local user karaf | JasperListener                   | 152 - org.apache.juli.extras - 7.0.32.v201211081135 | Couldn't initialize Jasper
java.lang.ClassNotFoundException: org.apache.jasper.compiler.JspRuntimeContext

In the log file after installation. We may need to track that down, but I do not believe this affects the suitability of this patch.

Patch-set: 5
Label: Verified=0