Update patch set 4
Patch Set 4:

Hi Tomas,
The CorsFilter is currently a fragment to the catalina bundle. It doesn't have to be. It could be made a Servlet/3.0 based Filter with annotations which reflect the default configuration (in which case it doesn't have to be defined in the web.xml). So, if a web bundle depends on the CorsFilter bundle, it will automatically have the filter configured. I have not tried this myself in an OSGi environment. So, I can't confirm it will work. IMO this would be a proper solution.

Lacking that, you can define the filter definition in the global web.xml. Then any webapp which needs the filter, can define the filter-mapping in its respective web.xml. This approach doesn't entirely get rid of CorsFilter in app specific web.xml. But, it ensures that the CorsFilter is configured consistently across all apps. Moreover developers understand web.xml well and filter-mapping belongs to the web.xml :)

I am not sure I follow the dynamic aspect of WABs. They have to be configured in cors-config.xml to pick up the CorsFilter.

Patch-set: 4