Update patch set 4
Patch Set 4:

Prasanth, 

I was hoping your solution would be based on tomcat-web.xml That was my reference to docBase because I believe this not doable in OSGi. The point of this change is to allow distributions and devops to customize security filters without touching upstream applications. 

This solution:

* does not introduce dependency of WABs on certain filters.

* does not force to use maven when customizing security rules

* allows reviewing and changing security rules from one place

> developers understand web.xml well

Controller developers have no visibility in how security filters should be configured because that is deployment specific information. Proposed solution is mimicking  tomcat-web.xml adding just one optional node <filter-template> that was added for readability.

Patch-set: 4