[DO NOT MERGE]
This is an extenstion to the existing wiki page on md-sal ping tutorial.
It exposes the use of Future<> with the rpc api generated by yang-tools,
to make the ping service non-blocking; using the uris:

   .../ping/async/start/{ipAddress}
   .../ping/async/get/{ipAddress}
   .../ping/async/stop/{ipAddress}

If this becomes interesting enough, I will update the link https://wiki.opendaylight.org/view/Ping
with the stepping stones.

Example:

for x in 127.0.0.1 128.0.0.1 192.168.1.1 ;
   do curl --user "admin":"admin" -X PUT  http://localhost:8080/controller/nb/v2/ping/async/start/${x} ; echo ; done

while : ; do for x in 127.0.0.1 128.0.0.1 192.168.1.1 ;
   do curl --user "admin":"admin" -X PUT  http://localhost:8080/controller/nb/v2/ping/async/get/${x} ; echo ;
done ; sleep 2 ; done

for x in 127.0.0.1 128.0.0.1 192.168.1.1 ;
   do curl --user "admin":"admin" -X PUT  http://localhost:8080/controller/nb/v2/ping/async/stop/${x} ; echo ; done

[DO NOT MERGE]

Change-Id: Ibfaee34ace0435cac39006a7e9c6577df2bf71a8
Signed-off-by: Flavio Fernandes <ffernand@redhat.com>