Bug 992 - Fix broken netconf xml serialization.
Empty commit and restart triggered xml deserialization errors
in netconf. The problem is in type and name elements - they
should not be prefixed with 'prefix:'.

This fix triggers other bugs in md-sal modules, tracked
as dependencies of this bug. To test proper (de)serialization
with this fix, 01-netconf-connector.xml and 02-clustering.xml
must be delted.

Change-Id: I54d6ab23ec27ccbed924bf6abff95736f49dd5a3
Signed-off-by: Tomas Olvecky <tolvecky@cisco.com>