Update patch set 1
Patch Set 1: Code-Review-1

On May 14, 2014, at 7:02 PM, Rob Adams <readams@readams.net> wrote:

I'm not sure I totally understand this, but it appears as though there's a fixed global currResult object.  This is very broken and almost never what you would want for an asynchronous API (or really any API) because it assumes only one user of the API at a time.

Implementing correct asynchronous semantics requires:
(1) event-driven I/O if you want to scale past single to low double-digit numbers of clients
(2) not using anything servlet based, which precludes event-driven I/O
(3) a mechanism for asynchronous dispatch of results; could be long polling or notifications, but cannot block threads.

Unfortunately I can't really see any way to implement anything asynchronous using the RPCs the way they're defined without starting lots of threads that can sit and wait on futures, which is obviously not a good option.

-Rob


You are right, Rob! I totally see what you mean: what is the point of making it ‘look’ async if internally it cannot handle more than one
ongoing ping, right? Thinking in those terms -- together with the fact that this should be a working example folks may use as boiler
plate -- it is very important to implement it right. Or at least, not advertise bad practices. :)

I did not know there are upcoming changes in the restconf. So, I’m a little hesitant that making it right may make the example a bit
complex, which would jeopardize the real purpose of the ping tutorial. Still, I want to make it right… so let me internalize all the feedback
into an implementation that holds water. For sake of completeness.

Thank you so much for your time looking at this and the feedback!

— flavio

Patch-set: 1