Update patch set 4
Patch Set 4:

Hi Tomas,
In my change, I have shown that you can tie in configuration with the bundle and it will get applied. And the dependent module doesn't have to declare the configuration. 

If you are trying to solve other issues (beyond the CORS filter as stated in the change description or the bug), then I am unaware of those requirements. The configuration in cors-config.xml is independent of the CORS filter fragment. If tomorrow someone moves the fragment (as I did in the change), the filter-valve will start failing. As I see it, you are trying to drive the CORS filter from a valve. Why not make the CORS filter into a valve?

I am not sure I am convinced that this is making it any more easier:

[TO] does not introduce dependency of WABs on certain filters.

Not entirely true. The dependency still exists - its been moved from the WAB to cors-config.xml. Also, you have an undeclared dependencies from filter-valve on classes which are in the cors-config.xml. Also, the context-roots need to be declared in cors-config.xml I am not sure if this is any better state. 
 
[TO] does not force to use maven when customizing security rules
[TO] allows reviewing and changing security rules from one place
 
These seem to suggest you are trying to solve something beyond the scope of CORS. Do you expect security rules(?) to be implemented as filters ? Do you expect to add other aspects of web.xml to core-config.xml ?

Patch-set: 4